test: split up test files to avoid worker overlap
