+++ tests/e2e/main.spec.js
@@ -1,4 +1,3 @@
-import { expect, test } from "@playwright/test";
 // import {
 //   clickMenuItemById,
 //   ipcMainCallFirstListener,
@@ -6,71 +5,8 @@ import { expect, test } from "@playwright/test";
 //   ipcMainInvokeHandler,
 //   ipcRendererInvoke
 // } from "electron-playwright-helpers";
-import { modVApp } from "./pageObjectModel";
+
 // import jimp from "jimp";
-
-test("renders the main window", async () => {
-  const { page } = modVApp;
-
-  const title = await page.title();
-  expect(title).toBe("modV");
-});
-
-test("default group is focused by default", async () => {
-  const { page } = modVApp;
-
-  const firstGroupEl = page.locator(".group").first();
-  const firstGroupElId = (await firstGroupEl.getAttribute("id")).replace(
-    "group-",
-    ""
-  );
-  const focusIndicator = firstGroupEl.locator(".group__focusIndicator");
-
-  await expect(focusIndicator).toHaveCount(1);
-
-  const { focus } = await modVApp.evaluateUIState();
-  expect(focus.id).toBe(firstGroupElId);
-});
-
-test("renderers are registered", async () => {
-  const expectedRenderers = ["2d", "isf", "shader", "three"];
-
-  const { renderers } = await modVApp.evaluateMainState();
-
-  expect(expectedRenderers.sort()).toEqual(Object.keys(renderers).sort());
-});
-
-test("search gallery and add module to group", async () => {
-  const { page } = modVApp;
-
-  const searchInput = await page.locator("input.gallery-search");
-  await searchInput.click();
-
-  await searchInput.fill("ball");
-
-  // Double click c:nth-child(2) > .smooth-dnd-container > div > .gallery-item > canvas >> nth=0
-  await page
-    .locator("c:nth-child(2) > .smooth-dnd-container > div > .gallery-item ", {
-      hasText: "Ball"
-    })
-    .first()
-    .dblclick({
-      position: {
-        x: 98,
-        y: 27
-      }
-    });
-
-  await page.waitForSelector(".group__modules .active-module");
-  const activeModules = await page.locator(".group__modules .active-module");
-  expect(await activeModules.count()).toBe(1);
-
-  const activeModule = await activeModules
-    .first()
-    .locator(".active-module__name");
-  const activeModuleTitle = await activeModule.textContent();
-  expect(activeModuleTitle.trim()).toBe("Ball");
-});
 
 // test("trigger IPC listener via main process", async () => {
 //   electronApp.evaluate(({ ipcMain }) => {
