+++ tests/e2e/spec/general/rendersTheMainWindow.spec.js
@@ -1,0 +1,9 @@
+import { expect, test } from "@playwright/test";
+import { modVApp } from "../../pageObjectModel";
+
+test("renders the main window", async () => {
+  const { page } = modVApp;
+
+  const title = await page.title();
+  expect(title).toBe("modV");
+});
