+++ tests/e2e/spec/groups/newGroupButtonCreatesANewGroup.spec.js
@@ -1,0 +1,14 @@
+import { expect, test } from "@playwright/test";
+import { modVApp } from "../../pageObjectModel";
+
+test("new group button creates a new group", async () => {
+  const { length: groupsLength } = (await modVApp.groups.mainState()).groups;
+
+  await modVApp.groups.newGroupButton.click();
+
+  await expect(modVApp.groups.elements).toHaveCount(groupsLength);
+
+  await modVApp.checkWorkerAndMainState(state =>
+    expect(state.groups.groups.length).toBe(groupsLength + 1)
+  );
+});
